[RAPTOR-3088] Hide error server output (similarly to pred server output) (#73)
* Hide error server output (similarly to pred server output)

* Undo log level setting

* Fix test

* update changelog
