+++ custom_model_runner/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.1.3] - in progress
+### Added
+- error server is started in case of imports/pipeline failures
 ### Changed
 - updated Java predictor dependencies
